Swap "embarrassing" for "complementary"
In an effort to be more inclusive and build an internet we all want to be a part of. Perhaps we could, in the learning stage, create environments that uplift and inspire. I've been told that this is a joke to bring humor to dry content and I really appreciate that about Odin in general. 

However, in this case, I can't imagine posting anything embarrassing anonymously about another person that wouldn't be potentially hurtful. I've been told this hasn't been an issue in the past, but just because it hasn't been an issue in the past, doesn't mean it's not an issue.